refactor proposal client consent router
