+++ src/app/routers/proposal_workflow_decisions.py
@@ -72,35 +72,3 @@ async def approve_compliance(
         idempotency_key=idempotency_key,
         correlation_id=correlation_id,
     )
-
-
-@router.post(
-    "/{proposal_id}/record-client-consent",
-    response_model=ProposalStateTransitionEnvelopeResponse,
-    summary="Record Proposal Client Consent",
-    description="Records client consent for a proposal that has completed internal approvals.",
-)
-async def record_client_consent(
-    request: ProposalApprovalActionRequest,
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-    idempotency_key: str = Header(
-        alias="Idempotency-Key",
-        description="Caller-supplied idempotency key for client-consent recording requests.",
-        examples=["idem-client-consent-1"],
-    ),
-) -> ProposalStateTransitionEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.record_client_consent(
-        proposal_id=proposal_id,
-        actor_id=request.actor_id,
-        expected_state=request.expected_state,
-        details=request.details,
-        related_version_no=request.related_version_no,
-        idempotency_key=idempotency_key,
-        correlation_id=correlation_id,
-    )
